Fix enum printing via magic_enum
- Enums were not recognized as StreamInsertable concepts.
- This was fixed by moving the PrintHelper functions to the
corresponding parameter container template file and including the
corresponding enum utilities file.
- A dedicated unit test was added checking the stream-insertability.
